% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/flatDataFrame.R
\name{flatDataframe}
\alias{flatDataframe}
\title{Return a "flat" data.frame where colums that are of class data.frame are integrated into the parent data.frame}
\usage{
flatDataframe(x)
}
\arguments{
\item{x}{a data.frame}
}
\value{
a data.frame
}
\description{
Return a "flat" data.frame where colums that are of class data.frame are integrated into the parent data.frame
}
\examples{
a<-data.frame(v1=seq(1:10),v2=letters[1:10])
a$complex<-a
flatDataframe(a)
}

