% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/gcalendar-package.R
\docType{package}
\name{gcalendar}
\alias{gcalendar}
\alias{gcalendar-package}
\title{gcalendar}
\description{
A Google Calendars API client for R
}
\details{
R6 classes for querying the Google Calendars API
}
\keyword{package}

